refactor(types): centralize autobahn error sentinels
Move BlockDB and data-layer errors into types/errors.go and drop the
data package aliases so callers check types.ErrPruned/ErrNotFound/etc.
