+++ sei-tendermint/rpc/coretypes/responses.go
@@ -34,7 +34,7 @@ var (
 // when the caller has the active lower bound — e.g. CometBFT's
 // BlockStore.Base() in env.getHeight. Pass None when the caller only knows
 // the height was unavailable but not what the bound is — e.g. the Autobahn
-// path where data.GlobalBlock returns data.ErrPruned and the bound
+// path where data.GlobalBlock returns types.ErrPruned and the bound
 // (BlockDB's prune watermark) is internal to the store.
 //
 // Centralizing this format means both paths produce identical error
